making Navbar functional, getting auth working
