+++ client/src/ui/Navbar.tsx
@@ -1,9 +1,11 @@
-import React, { Component, useEffect, useState } from "react";
+import React from "react";
 import "./css/Navbar.css";
 import SearchBar from "./SearchBar.jsx";
 import { Session } from "../session-store";
 import ProfileDropdownNavBar from "./ProfileDropdownNavBar.jsx";
 import { useAuthOptionalLogin } from "../auth/auth_utils";
+import { randomPicture } from "../util/profile_picture";
+import { useLocation } from "react-router-dom";
 
 /*
   Navbar Component. Short description if needed.
@@ -17,8 +19,6 @@ import { useAuthOptionalLogin } from "../auth/auth_utils";
   component tree, and any inportant information it accesses or modifies.
   Include the route for View components.
 */
-
-const [isLoggedIn, token, signIn, signOut] = useAuthOptionalLogin();
 
 const sunset_start_times = [
   17.0, 17.5, 18, 19.5, 20, 20.5, 20.5, 19.5, 18.5, 18, 16.5, 16.5,
@@ -54,70 +54,61 @@ if (time_of_day < 6 || time_of_day >= sunset_end_times[month]) {
   dayclass = "sunset";
 }
 
-const profilePictures = [
-  "/profile_bear/profile_bear_dark_blue.svg",
-  "/profile_bear/profile_bear_light_blue.svg",
-  "/profile_bear/profile_bear_light_pink.svg",
-  "/profile_bear/profile_bear_mint.png",
-  "/profile_bear/profile_bear_orange.svg",
-  "/profile_bear/profile_bear_purple.svg",
-  "/profile_bear/profile_bear_red.svg",
-  "/profile_bear/profile_bear_yellow.svg",
-];
-
-function randomPicture() {
-  return profilePictures[Math.floor(Math.random() * profilePictures.length)];
-}
-
 const profilePicture = randomPicture();
 
-function displayButton() {
-  const token = Session.get("token");
-  if (token) {
-    return (
-      <ProfileDropdownNavBar
-        imgSrc={`${String(profilePicture)}`}
-        isLoggedIn={token}
-        signOut={signOut}
-      />
-    );
-  } else {
-    return (
-      <button
-        type="button"
-        className="btn btn-light sign-in-button"
-        onClick={() => {
-          signIn("home");
-        }}
-      >
-        Sign In
-      </button>
-    );
-  }
-}
+type NavbarProps = {
+  userInput: string;
+};
 
-export default class Navbar extends Component<{ readonly userInput: string }> {
-  render() {
-    return (
-      <div className="custom-navbar">
-        <div className="logo-container">
-          <a className="" href="/">
-            <img
-              src="/logo.svg"
-              className="img-fluid scale-logo-navbar"
-              alt="CU Reviews Logo"
-            />
-          </a>
-        </div>
-        <div className="col navbar-searchbar-container">
-          <SearchBar
-            userInput={this.props.userInput}
-            contrastingResultsBackground={true}
-            isInNavbar={true}
-          />
-        </div>
-        {displayButton()}
-      </div>
-    );
+export default function Navbar({ userInput }: NavbarProps) {
+
+  const [isLoggedIn, token, signIn, signOut] = useAuthOptionalLogin();
+  const location = useLocation();
+
+  function displayButton() {
+    const token = Session.get("token");
+    if (token) {
+      return (
+        <ProfileDropdownNavBar
+          imgSrc={`${String(profilePicture)}`}
+          isLoggedIn={token}
+          signOut={signOut}
+        />
+      );
+    } else {
+      return (
+        <button
+          type="button"
+          className="btn btn-light sign-in-button"
+          onClick={() => {
+            signIn("path:" + location.pathname);
+          }}
+        >
+          Sign In
+        </button>
+      );
+    }
   }
+
+  return (
+    <div className="custom-navbar">
+      <div className="logo-container">
+        <a className="" href="/">
+          <img
+            src="/logo.svg"
+            className="img-fluid scale-logo-navbar"
+            alt="CU Reviews Logo"
+          />
+        </a>
+      </div>
+      <div className="col navbar-searchbar-container">
+        <SearchBar
+          userInput={userInput}
+          contrastingResultsBackground={true}
+          isInNavbar={true}
+        />
+      </div>
+      {displayButton()}
+    </div>
+  );
 }
